update download functions
